requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.30.0
google-genai>=0.3.0
